requirements.txt
asgiref==3.9.1
Django==4.2.24
djangorestframework==3.16.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
